+++ Stellar.xcodeproj/xcuserdata/elliotknight.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -17,12 +17,12 @@
 		<key>StellarNotificationServiceExtension.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>3</integer>
+			<integer>4</integer>
 		</dict>
 		<key>StellarStickers.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>3</integer>
 		</dict>
 		<key>StellarTests.xcscheme_^#shared#^_</key>
 		<dict>
